requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
coverage==7.6.10
cryptography==45.0.0
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.7
django-sslserver==0.22
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
Fibo==1.0.0
firebase-admin==6.9.0
Flask==3.1.0
fonttools==4.58.2
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.6.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==20.1.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
instaloader==4.14.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
msgpack==1.1.1
multidict==6.4.3
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.5
openai==0.28.0
orjson==3.11.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipdeptree==2.26.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.2
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pypdf==5.5.0
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.20
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==308
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
sendgrid==6.11.0
sentry-sdk==2.33.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
starlette==0.47.2
stripe==11.5.0
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20241230
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.3.0
uvicorn==0.35.0
waitress==3.0.2
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
xhtml2pdf==0.2.17
yarl==1.20.0
